refactor: utility functions removed to a different directory.
